Make remote executor more generic
Modify the remote executor to take a remote executor string instead of the
target IP address and port.  The command executed on the host is the
remote executor string (e.g., "my_exec --target-id 1234") plus the command
to be executed on the target (e.g., "echo hello") as a single additional
argument.

This allows the remote executor to use commands other than SSH to actually
run the command on the target.  This commit only implements the
lower-level support and doesn't expose this ability to the user.