Allow more than SCP for getting remote files
Previously, SCP was always used to get files from a remote target.  This
commit adds the ability to use others commands specified by a getter
string (e.g., "my_getter --target-id 1234 <FILE> <DEST>").  To get a
file from the target to a destination directory on the host,
testrunner-lite replaces <FILE> with the path to the file and <DEST> with
the path to the destination directory, then executes the command.

This commit only implements the lower-level support and doesn't expose
this ability to the user.